+++ src/main/java/TaskList.java
@@ -4,14 +4,11 @@
  * @author JamesLiuZX
  * AY2223-S2 CS2103T
  */
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 
-public class DukeList {
+public class TaskList {
     //Variables are kept protected and accessed only through internal getter and setter methods.
     protected static final int size = 100;
     protected ArrayList<Task> records;
@@ -21,7 +18,7 @@ public class DukeList {
     private DatabaseWriter dw = new DatabaseWriter(path);
 
     //Constructor
-    public DukeList() {
+    public TaskList() {
         records = new ArrayList<>(size);
     }
 
@@ -33,7 +30,7 @@ public class DukeList {
         Task task = records.get(index);
         task.markDone();
         dw.setDone(index);
-        System.out.println(Duke.format("Nice! I've marked this task as done:\n" + task.toString()));
+        System.out.println(Ui.format("Nice! I've marked this task as done:\n" + task.toString()));
         return;
     }
 
@@ -45,7 +42,7 @@ public class DukeList {
         Task task = records.get(index);
         task.markUndone();
         dw.setUndone(index);
-        System.out.println(Duke.format("OK, I've marked this task as not done yet:\n\n" + task.toString()));
+        System.out.println(Ui.format("OK, I've marked this task as not done yet:\n\n" + task.toString()));
         return;
     }
 
@@ -53,14 +50,14 @@ public class DukeList {
         Task t = new Task(record);
         records.add(t);
         dw.addToDb(t);
-        System.out.println(Duke.format("added: " + t));
+        System.out.println(Ui.format("added: " + t));
     }
 
     public void insertToDo(String name) {
         TaskToDo t = new TaskToDo(name);
         records.add(t);
         dw.addToDb(t);
-        System.out.println(Duke.format("Got it. I've added this task:\n" + t.toString()));
+        System.out.println(Ui.format("Got it. I've added this task:\n" + t.toString()));
     }
 
     public void insertToDo(String name, boolean isInitial) {
@@ -72,7 +69,7 @@ public class DukeList {
         TaskDeadline d = new TaskDeadline(name, time);
         records.add(d);
         dw.addToDb(d);
-        System.out.println(Duke.format("Got it. I've added this task:\n" + d.toString()));
+        System.out.println(Ui.format("Got it. I've added this task:\n" + d.toString()));
     }
     public void insertDeadline(String name, String time, boolean isInitial) {
         TaskDeadline d = new TaskDeadline(name, time);
@@ -86,7 +83,7 @@ public class DukeList {
         TaskEvent e = new TaskEvent(name, from, to);
         records.add(e);
         dw.addToDb(e);
-        System.out.println(Duke.format("Got it. I've added this task:\n" + e.toString()));
+        System.out.println(Ui.format("Got it. I've added this task:\n" + e.toString()));
     }
     public void insertEvent(String name, String time, boolean isInitial) {
         String[] times = time.split("/", 2);
@@ -104,7 +101,7 @@ public class DukeList {
         Task t = records.get(index);
         records.remove(index);
         dw.removeFromDb(index);
-        System.out.println(Duke.format("Noted. I've removed this task:\n" + t.toString()));
+        System.out.println(Ui.format("Noted. I've removed this task:\n" + t.toString()));
     }
 
     @Override
@@ -116,7 +113,7 @@ public class DukeList {
         for (int i = 0; i < records.size(); i++) {
             output += String.format("%s. %s\n", i+1, records.get(i));
         }
-        return Duke.format(output);
+        return Ui.format(output);
     }
